imp: stream流只传文字
