+++ server/utils/index.ts
@@ -10,4 +10,32 @@ function getSystemMessage (): ChatCompletionRequestMessage[] {
   return [currentTime, ...systemMessages];
 }
 
-export { getSystemMessage };
+/**
+ * 解析stream流的字符串
+ */
+function parseStreamText(data: string) {
+  const dataList = data?.split('\n')?.filter((l) => l !== '');
+
+  const result = { role: 'assistant', content: '', stop: false };
+
+  dataList.forEach((l) => {
+    // 移除"data: "前缀
+    const jsonStr = l.replace('data: ', '');
+
+    if (jsonStr === '[DONE]') {
+      result.stop = true;
+    } else {
+      // 将JSON字符串转换为JavaScript对象
+      const jsonObj = JSON.parse(jsonStr);
+      const delta = jsonObj.choices[0].delta as ChatCompletionRequestMessage;
+      if (delta.role) result.role = delta.role;
+      else if (delta.content) {
+        result.content = `${result.content}${delta.content}`;
+      }
+    }
+  });
+
+  return result;
+}
+
+export { getSystemMessage, parseStreamText };
