+++ src/utils/index.ts
@@ -27,39 +27,15 @@ function getCachePrompt (conversation: Conversation[], curValue: string): Messag
   }
 }
 
-/**
- * 解析stream流的字符串
- */
-function parseStreamText(data: string) {
-  const dataList = data?.split('\n')?.filter((l) => l !== '');
-
-  const result = { role: 'assistant', content: '', stop: false };
-
-  dataList.forEach((l) => {
-    // 移除"data: "前缀
-    const jsonStr = l.replace('data: ', '');
-
-    if (jsonStr === '[DONE]') {
-      result.stop = true;
-    } else {
-      // 将JSON字符串转换为JavaScript对象
-      const jsonObj = JSON.parse(jsonStr);
-      const delta = jsonObj.choices[0].delta as Messages;
-      if (delta.role) result.role = delta.role;
-      else if (delta.content) {
-        result.content = `${result.content}${delta.content}`;
-      }
-    }
-  });
-
-  const matches = result.content.match(/```/g);
+function parseMarkdown(chunk: string): string {
+  let text = chunk;
+  const matches = chunk.match(/```/g);
   const count = matches ? matches.length : 0;
   if (count % 2 !== 0) {
     // 如果计数为奇数，说明```没有成对，因此在字符串末尾添加```
-    result.content += '\n```';
+    text += '\n```';
   }
-
-  return result;
+  return text;
 }
 
-export { getCachePrompt, parseStreamText };
+export { getCachePrompt, parseMarkdown };
